src/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
redis==5.0.1
openai==1.10.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
